CI passing for P256 (HW wallet restricted to K256)
