+++ src/curve.rs
@@ -1,11 +1,7 @@
 //! Elliptic Curve abstraction
 use crate::{errors::Result, k256::K256, p256::P256};
-use generic_array::GenericArray;
 use hmac::digest::core_api::CoreWrapper;
-use k256::{
-    elliptic_curve::{scalar::IsHigh, Field, PrimeField},
-    FieldBytes, Scalar as K256_Scalar,
-};
+use k256::FieldBytes;
 use libpaillier::unknown_order::BigNumber;
 use serde::{Deserialize, Serialize};
 use sha3::Keccak256Core;
@@ -202,76 +198,13 @@ pub trait ScalarTrait:
     fn invert(&self) -> Option<Self>;
 }
 
-impl ScalarTrait for K256_Scalar {
-    fn zero() -> Self {
-        K256_Scalar::ZERO
-    }
-
-    fn one() -> Self {
-        K256_Scalar::ONE
-    }
-
-    fn convert_from_u128(x: u128) -> Self {
-        K256_Scalar::from_u128(x)
-    }
-
-    fn add(&self, other: &Self) -> Self {
-        k256::Scalar::add(self, other)
-    }
-
-    fn sub(&self, other: &Self) -> Self {
-        k256::Scalar::sub(self, other)
-    }
-
-    fn negate(&self) -> Self {
-        k256::Scalar::negate(self)
-    }
-
-    fn mul(&self, other: &Self) -> Self {
-        k256::Scalar::mul(self, other)
-    }
-
-    fn mul_bignum(&self, other: &BigNumber) -> Self {
-        // use bn_to_scalar to convert other to a scalar
-        let bn_scalar: Self = <K256 as CurveTrait>::bn_to_scalar(other).unwrap();
-        k256::Scalar::mul(self, &bn_scalar)
-    }
-
-    fn is_high(&self) -> bool {
-        <k256::Scalar as IsHigh>::is_high(self).into()
-    }
-
-    fn random() -> Self {
-        let rng = rand::thread_rng();
-        <K256_Scalar as Field>::random(rng)
-    }
-
-    fn to_bytes(&self) -> Vec<u8> {
-        K256_Scalar::to_bytes(self).to_vec()
-    }
-
-    fn from_bytes(bytes: &[u8]) -> Option<Self> {
-        <K256_Scalar as PrimeField>::from_repr(GenericArray::clone_from_slice(bytes)).into()
-    }
-
-    fn from_repr(bytes: Vec<u8>) -> Self {
-        <K256_Scalar as PrimeField>::from_repr(GenericArray::clone_from_slice(&bytes)).unwrap()
-    }
-
-    fn modulus(&self) -> BigNumber {
-        BigNumber::from_slice(<K256_Scalar as PrimeField>::MODULUS)
-    }
-
-    fn invert(&self) -> Option<Self> {
-        K256_Scalar::invert(self).into()
-    }
-}
-
 /// Default curve type.
-pub type TestCurve = K256;
+//pub type TestCurve = K256;
+pub type TestCurve = P256;
 
 /// Default scalar type.
-pub type TestScalar = K256_Scalar;
+//pub type TestScalar = K256_Scalar;
+pub type TestScalar = p256::Scalar;
 
 /// Default signature type.
 pub type TestSignature = k256::ecdsa::Signature;
